redux works, components uses store now
