+++ keymaps/atom-solidity.cson
@@ -11,5 +11,5 @@
     'ctrl-alt-e': 'eth-interface:activate'
     'ctrl-alt-c': 'eth-interface:compile'
     'ctrl-alt-v': 'eth-interface:toggle'
-    "atom-text-editor[data-grammar~=\"solidity\"]:not([mini])":
+    'atom-text-editor[data-grammar~=\"solidity\"]:not([mini])':
         'ctrl-alt-e': 'eth-interface:activate'
