+++ lib/web3/view.js
@@ -2,11 +2,11 @@
 
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { Tab, Tabs, TabList, TabPanel } from 'react-tabs'
 import createReactClass from 'create-react-class'
 import ReactUpdate from 'react-addons-update'
 import ReactJson from 'react-json-view'
 import Web3Helpers from './methods'
+import TabView from '../components/TabView'
 
 export default class View {
 	constructor(store, web3) {
@@ -15,9 +15,6 @@ export default class View {
 		this.web3 = web3;
 		this.store = store;
 		this.helpers = new Web3Helpers(this.web3);
-	}
-	setCompiled(compiled) {
-		this.compiled = compiled;
 	}
 	viewErrors(errors) {
 		let errorViews = createReactClass({
@@ -190,283 +187,7 @@ export default class View {
 		ReactDOM.render(React.createElement(compileButton, null), document.getElementById('compile_btn'));
 	}
 	async createTabView() {
-		let workspaceElement = atom.views.getView(atom.workspace);
-		const that = this;
-		const { contract } = this.store.getState();
-		const { compiled } = contract;
-		console.log(compiled);
-		let params = [];
-		class InputsForm extends React.Component {
-			constructor(props) {
-				super(props);
-				this._handleChange = this._handleChange.bind(this);
-			}
-			_handleChange(event) {
-				params[event.target.id] = { value: event.target.value, type: event.target.placeholder };
-			}
-			render() {
-				const { contractName, abi } = this.props;
-				return (
-					<div id={contractName + '_inputs'}>
-						{
-							abi.type === "constructor" &&
-							abi.inputs.map((input, i) => {
-								return (
-									<form key={i}>
-										<button class="input text-subtle">{ input.name }</button>
-										<input id={i} type="text" class="inputs" placeholder={input.type} onChange={this._handleChange}></input>
-									</form>
-								);
-							})
-						}
-					</div>
-				);
-			}
-		};
-		class GasInput extends React.Component {
-			constructor(props) {
-				super(props);
-				this.state = {
-					gas: props.gas
-				};
-			}
-			componentWillReceiveProps(nextProps) {
-				const { gas } = nextProps;
-				this.setState({ gas });
-			}
-			render() {
-				const { contractName } = this.props;
-				return (
-					<form class="gas-estimate-form">
-						<button class="input text-subtle">Gas</button>
-						<input
-							id={contractName + '_gas'}
-							type="number"
-							class="inputs"
-							value={this.state.gas}
-							onChange={this.props.onChange}>
-						</input>
-					</form>
-				);
-			}
-		};
-		class CreateButton extends React.Component {
-			constructor(props) {
-				super(props);
-				this.state = {
-					constructorParams: undefined
-				}
-				this._handleSubmit = this._handleSubmit.bind(this);
-			}
-			async componentDidMount() {
-				const { abi } = this.props;
-				inputs = [];
-				for (abiObj in abi) {
-					if (abi[abiObj].type === "constructor" && abi[abiObj].inputs.length > 0) {
-						inputs = abi[abiObj].inputs;
-					}
-				}
-				this.setState({ constructorParams: inputs });
-			}
-			async _handleSubmit() {
-				console.log("%c Submit handle... ", 'background: rgba(36, 194, 203, 0.3); color: #EF525B');
-				try {
-					const { abi, bytecode, contractName, gas } = this.props;
-					const { constructorParams } = this.state;
-
-					const contract = await that.helpers.create(that.coinbase, that.password, abi, bytecode, params, contractName, gas);
-					that.setExecutionView(contractName, abi, bytecode, constructorParams, params, contract);
-					const contractInstance = await that.helpers.deploy(contract, params);
-					contractInstance.on('address', address => {
-						contract.options.address = address;
-						that.setExecutionView(contractName, abi, bytecode, constructorParams, params, contract);
-					});
-					contractInstance.on('transactionHash', transactionHash => {
-						contract.transactionHash = transactionHash;
-						that.setExecutionView(contractName, abi, bytecode, constructorParams, params, contract);
-					});
-					contractInstance.on('error', error => {
-						that.helpers.showPanelError(error);
-					});
-				} catch(e) {
-					that.helpers.showPanelError(e);
-				}
-			}
-			render() {
-				const { contractName } = this.props;
-				return (
-					<form onSubmit={this._handleSubmit}>
-						<input
-							type="submit"
-							value="Create"
-							ref={contractName}
-							class="btn btn-primary inline-block-tight">
-						</input>
-					</form>
-				);
-			}
-		};
-		class CompiledContract extends React.Component {
-			constructor(props) {
-				super(props);
-				this.state = {
-					estimatedGas: 4700000
-				}
-				this._handleGasChange = this._handleGasChange.bind(this);
-			}
-			async componentDidMount() {
-				const { bytecode } = this.props;
-				const gas = await that.helpers.getGasEstimate(that.coinbase, bytecode);
-				this.setState({ estimatedGas: gas });
-			}
-			_handleGasChange(event) {
-				this.setState({ estimatedGas: event.target.value });
-			}
-			render() {
-				const { contractName, bytecode, ContractABI, index } = this.props;
-				const { estimatedGas } = this.state;
-				return (
-					<div class="contract-content" key={index}>
-						<span class="contract-name inline-block highlight-success">{ contractName }</span>
-						<div class="byte-code">
-							<pre class="large-code">{ JSON.stringify(bytecode) }</pre>
-						</div>
-						<div class="abi-definition">
-							<pre class="large-code">{ JSON.stringify(ContractABI) }</pre>
-						</div>
-						{
-							ContractABI.map((abi) => {
-								return <InputsForm contractName={contractName} abi={abi} />
-							})
-						}
-						<GasInput contractName={contractName} gas={estimatedGas} onChange={this._handleGasChange} />
-						<CreateButton
-							contractName={contractName}
-							bytecode={bytecode}
-							abi={ContractABI}
-							gas={estimatedGas}
-						/>
-					</div>
-				);
-			}
-		};
-		class Contracts extends React.Component {
-			constructor(props) {
-				super(props);
-				const { contract } = that.store.getState();
-				this.state = {
-					compiled: contract.compiled
-				}
-			}
-			render() {
-				const { compiled } = this.state;
-				if(compiled) {
-					return (
-						<div id="compiled-code" class="compiled-code">
-							{
-								Object.keys(compiled.contracts).map((contractName, index) => {
-									const bytecode = compiled.contracts[contractName].bytecode;
-									const ContractABI = JSON.parse(compiled.contracts[contractName]["interface"]);
-									return (
-										<div id={contractName} class="contract-container">
-											<CompiledContract
-												contractName={contractName}
-												bytecode={bytecode}
-												ContractABI={ContractABI}
-												index={index}
-											/>
-										</div>
-									)
-								})
-							}
-						</div>
-					);
-				}
-				return (
-					<div id="compiled-code" class="compiled-code"></div>
-				);
-			}
-		}
-		class TabView extends React.Component {
-			constructor(props) {
-				super(props);
-				this.state = {
-					txHash: undefined,
-					txAnalysis: undefined
-				}
-				this._handleTxHashChange = this._handleTxHashChange.bind(this);
-				this._handleTxHashSubmit = this._handleTxHashSubmit.bind(this);
-			}
-			_handleTxHashChange(event) {
-				this.setState({ txHash: event.target.value });
-			}
-			async _handleTxHashSubmit() {
-				const { txHash } = this.state;
-				try {
-					const txAnalysis = await that.helpers.getTxAnalysis(txHash);
-					this.setState({ txAnalysis });
-				} catch (e) {
-					console.log(e);
-				}
-			}
-			render() {
-				return (
-					<Tabs>
-						<TabList>
-							<div class="tab_btns">
-								<Tab>
-									<div class="btn copy-btn btn-primary">Contract</div>
-								</Tab>
-								<Tab>
-									<div class="btn copy-btn btn-primary">Transaction analyzer</div>
-								</Tab>
-							</div>
-						</TabList>
-
-						<TabPanel>
-							<Contracts />
-						</TabPanel>
-						<TabPanel>
-							<div class="block">
-								<form onSubmit={this._handleTxHashSubmit}>
-									<div class="inline-block">
-										<input type="text" name="txhash" value={this.state.txHash} onChange={this._handleTxHashChange} placeholder="Transaction hash" class="input-search" />
-									</div>
-									<div class="inline-block">
-										<input type="submit" value="Analyze" class="btn btn-success" />
-									</div>
-								</form>
-							</div>
-							{
-								(this.state.txAnalysis && this.state.txAnalysis.transaction) &&
-								<div class="block">
-									<h2 class="block highlight-info tx-header">Transaction</h2>
-									<ReactJson
-										src={this.state.txAnalysis.transaction}
-										theme="solarized"
-									 	displayDataTypes={false}
-										name={false}
-									/>
-								</div>
-							}
-							{
-								(this.state.txAnalysis && this.state.txAnalysis.transactionRecipt) &&
-								<div class="block">
-									<h2 class="block highlight-info tx-header">Transaction receipt</h2>
-									<ReactJson
-										src={this.state.txAnalysis.transactionRecipt}
-										theme="solarized"
-									 	displayDataTypes={false}
-										name={false}
-									/>
-								</div>
-							}
-						</TabPanel>
-					</Tabs>
-				);
-			}
-		}
-		ReactDOM.render(React.createElement(TabView), document.getElementById('tab_view'));
+		ReactDOM.render(<TabView store={this.store} helpers={this.helpers}/>, document.getElementById('tab_view'));
 	}
 	async viewCompiled(compiled) {
 		const self = this;
